fracture model

+++ geometry/node.h
@@ -43,8 +43,8 @@ struct icy::Node
         icy::Node* nd[2];
         float centerAngle; // angle from the node to the center of the adjacent element
         float angle0, angle1;
-        Eigen::Vector2f u_normalized, v_normalized, u_p, v_p;
-        Eigen::Vector2f t0, t1;
+        Eigen::Vector2d u_normalized, v_normalized, u_p, v_p;
+        Eigen::Vector2d t0, t1;
     };
 
     bool isBoundary;
